Force beam pattern vector to be float.

+++ beam_and_timing.py
@@ -107,7 +107,7 @@ class beam_and_timing:
                     exec(f.read())
                     f.close()
             else:
-                ppb_vect = np.array(filling_pattern_file) 
+                ppb_vect = np.float_(np.array(filling_pattern_file)) 
                 
             if flag_secodary_beam:
                 t=t_primary_beam
